fixed a double/float conflict bug. VTK writes all time values.

+++ src/core3d/cuda/tetmesh.cu
@@ -11,7 +11,7 @@ using namespace std;
 
 #define MIN(a,b) std::min(a,b)
 
-void TetMesh::init(float* pointlist, int numpoint, int*trilist, int numtri, int* tetlist, int numtet, int numattr, float* attrlist, bool verbose)
+void TetMesh::init(double* pointlist, int numpoint, int*trilist, int numtri, int* tetlist, int numtet, int numattr, double* attrlist, bool verbose)
 {
 
   vertices.resize(numpoint);
